test(android): split the committed-mode caret tests off their own file
`check-kotlin-loc.sh` caps a file at 150 lines and the suite had grown to
172. The three tests that turn on who owns the caret answer a different
question from the four about how the buffer is written, so they move
together into `CommittedSelectionOwnershipTest`, with the handler setup
they share as one helper.
