+++ platforms/android/ime/src/test/java/app/funput/funput/ime/editing/CommittedCompositionSessionTest.kt
@@ -26,25 +26,6 @@ class CommittedCompositionSessionTest {
     }
 
     @Test
-    fun `selection update from committed replacement keeps engine buffer`() {
-        val editor = CommittedEditor()
-        val session = testSession(ScriptedEngine(ArrayDeque(listOf("chào"))))
-        val handler = ImeKeyActionHandler(
-            composition = session,
-            editor = InputConnectionEditor(),
-            connection = { editor.proxy },
-            enterCommand = { ImeEditCommand.CommitText("\n") },
-        )
-        handler.start(renderMode = CompositionRenderMode.COMMITTED)
-
-        handler.onKeyAction(KeyAction.Input(keyId = "character-a", text = "a"))
-        handler.onSelectionChanged(newStart = 4, newEnd = 4, composingEnd = -1)
-
-        assertTrue(session.isComposing)
-        assertEquals("chào", editor.text)
-    }
-
-    @Test
     fun `boundary replacement rewrites committed word once`() {
         val editor = CommittedEditor()
         val session = testSession(
@@ -113,60 +94,5 @@ class CommittedCompositionSessionTest {
 
         assertEquals("dúng", editor.text)
         assertEquals("dung", engine.adopted)
-    }
-
-    @Test
-    fun `caret reported from inside our batch edit keeps the buffer`() {
-        val editor = CommittedEditor()
-        val session = testSession(ScriptedEngine(ArrayDeque(listOf("u", "uo", "uơ", "ươn"))))
-        val handler = ImeKeyActionHandler(
-            composition = session,
-            editor = InputConnectionEditor(),
-            connection = { editor.proxy },
-            enterCommand = { ImeEditCommand.CommitText("\n") },
-        )
-        handler.start(renderMode = CompositionRenderMode.COMMITTED)
-
-        handler.onKeyAction(KeyAction.Input(keyId = "character-u", text = "u"))
-        handler.onSelectionChanged(newStart = 1, newEnd = 1, composingEnd = -1)
-        handler.onKeyAction(KeyAction.Input(keyId = "character-o", text = "o"))
-        handler.onSelectionChanged(newStart = 2, newEnd = 2, composingEnd = -1)
-        // Replacing "uo" with "uơ" deletes before it commits; a host that ignores the
-        // batch reports the empty document in between before reporting the result.
-        handler.onKeyAction(KeyAction.Input(keyId = "character-w", text = "w"))
-        editor.textBeforeCursor = ""
-        handler.onSelectionChanged(newStart = 0, newEnd = 0, composingEnd = -1)
-        editor.textBeforeCursor = null
-        handler.onSelectionChanged(newStart = 2, newEnd = 2, composingEnd = -1)
-        handler.onKeyAction(KeyAction.Input(keyId = "character-n", text = "n"))
-
-        assertEquals("ươn", editor.text)
-        assertEquals("ươn", session.composingText)
-    }
-
-    @Test
-    fun `caret move after the batch settles releases the buffer`() {
-        val editor = CommittedEditor()
-        val session = testSession(ScriptedEngine(ArrayDeque(listOf("u", "uo", "uơ"))))
-        val handler = ImeKeyActionHandler(
-            composition = session,
-            editor = InputConnectionEditor(),
-            connection = { editor.proxy },
-            enterCommand = { ImeEditCommand.CommitText("\n") },
-        )
-        handler.start(renderMode = CompositionRenderMode.COMMITTED)
-
-        handler.onKeyAction(KeyAction.Input(keyId = "character-u", text = "u"))
-        handler.onSelectionChanged(newStart = 1, newEnd = 1, composingEnd = -1)
-        handler.onKeyAction(KeyAction.Input(keyId = "character-o", text = "o"))
-        handler.onSelectionChanged(newStart = 2, newEnd = 2, composingEnd = -1)
-        // The replacement settles first, which is what retires the caret we expect from
-        // inside the batch. The tap that follows is the user's, and has to be obeyed.
-        handler.onKeyAction(KeyAction.Input(keyId = "character-w", text = "w"))
-        handler.onSelectionChanged(newStart = 2, newEnd = 2, composingEnd = -1)
-        editor.textBeforeCursor = ""
-        handler.onSelectionChanged(newStart = 0, newEnd = 0, composingEnd = -1)
-
-        assertEquals("", session.composingText)
     }
 }
